ci: verification page updates
