+++ submissions/kivumbiGeorge/react/react-app/src/components/OrderSuccessful/success.css
@@ -37,7 +37,7 @@
   color: #5ea304;
 }
 
-.text-paragraph-top {
+.text-paragraph {
   font-family: "Poppins";
   font-style: normal;
   font-weight: 400;
@@ -47,16 +47,6 @@
   margin-top: -10px;
   width: 80%;
   text-align: center;
-}
-
-.text-paragraph-bottom {
-  font-family: "Poppins";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 30px;
-  color: #696969;
-  margin-top: -20px;
 }
 
 @media screen and (min-width: 320px) and (max-width: 768px) {
@@ -99,7 +89,7 @@
     font-size: 15px;
     line-height: 22px;
     margin-top: -10px;
-    width: 60%;
+    width: 70%;
   }
