+++ submissions/kivumbiGeorge/react/react-app/src/components/Verification/verification.css
@@ -161,6 +161,10 @@
     font-size: 16px;
 }
 
+.null-select {
+    display: none;
+}
+
 @media screen and (min-width: 320px) and (max-width: 768px) {
 
     .main-body {
