+++ tests/logfile.json
@@ -1,1 +1,1 @@
-{"numFailedTestSuites":0,"numFailedTests":0,"numPassedTestSuites":1,"numPassedTests":1,"numPendingTestSuites":0,"numPendingTests":0,"numRuntimeErrorTestSuites":0,"numTodoTests":0,"numTotalTestSuites":1,"numTotalTests":1,"openHandles":[],"snapshot":{"added":0,"didUpdate":false,"failure":false,"filesAdded":0,"filesRemoved":0,"filesRemovedList":[],"filesUnmatched":0,"filesUpdated":0,"matched":0,"total":0,"unchecked":0,"uncheckedKeysByFile":[],"unmatched":0,"updated":0},"startTime":1676119622806,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":[],"failureMessages":[],"fullName":"renders without crashing","location":null,"status":"passed","title":"renders without crashing"}],"endTime":1676119624481,"message":"","name":"/Users/gokivumbi/Documents/Projects/Else/Dufuna-CodeCamp-21/submissions/kivumbiGeorge/react/react-app/src/__tests__/App.test.js","startTime":1676119623425,"status":"passed","summary":""}],"wasInterrupted":false}
+{"numFailedTestSuites":0,"numFailedTests":0,"numPassedTestSuites":4,"numPassedTests":14,"numPendingTestSuites":0,"numPendingTests":0,"numRuntimeErrorTestSuites":0,"numTodoTests":0,"numTotalTestSuites":4,"numTotalTests":14,"openHandles":[],"snapshot":{"added":0,"didUpdate":false,"failure":false,"filesAdded":0,"filesRemoved":0,"filesRemovedList":[],"filesUnmatched":0,"filesUpdated":0,"matched":0,"total":0,"unchecked":0,"uncheckedKeysByFile":[],"unmatched":0,"updated":0},"startTime":1676120755457,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":[],"failureMessages":[],"fullName":"renders without crashing","location":null,"status":"passed","title":"renders without crashing"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: assert image amount is 1","location":null,"status":"passed","title":"test case: assert image amount is 1"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check order success title","location":null,"status":"passed","title":"test case: check order success title"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check order success description","location":null,"status":"passed","title":"test case: check order success description"}],"endTime":1676120756896,"message":"","name":"/Users/gokivumbi/Documents/Projects/Else/Dufuna-CodeCamp-21/submissions/kivumbiGeorge/react/react-app/src/__tests__/OrderSuccessful.test.js","startTime":1676120756164,"status":"passed","summary":""},{"assertionResults":[{"ancestorTitles":[],"failureMessages":[],"fullName":"renders without crashing","location":null,"status":"passed","title":"renders without crashing"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check service-title1","location":null,"status":"passed","title":"test case: check service-title1"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check input elements","location":null,"status":"passed","title":"test case: check input elements"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check select element","location":null,"status":"passed","title":"test case: check select element"}],"endTime":1676120756931,"message":"","name":"/Users/gokivumbi/Documents/Projects/Else/Dufuna-CodeCamp-21/submissions/kivumbiGeorge/react/react-app/src/__tests__/Verification.test.js","startTime":1676120756161,"status":"passed","summary":""},{"assertionResults":[{"ancestorTitles":[],"failureMessages":[],"fullName":"renders without crashing","location":null,"status":"passed","title":"renders without crashing"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check header title","location":null,"status":"passed","title":"test case: check header title"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check intial counter value","location":null,"status":"passed","title":"test case: check intial counter value"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: expected images","location":null,"status":"passed","title":"test case: expected images"},{"ancestorTitles":[],"failureMessages":[],"fullName":"test case: check submit button","location":null,"status":"passed","title":"test case: check submit button"}],"endTime":1676120756944,"message":"","name":"/Users/gokivumbi/Documents/Projects/Else/Dufuna-CodeCamp-21/submissions/kivumbiGeorge/react/react-app/src/__tests__/CartPage.test.js","startTime":1676120756161,"status":"passed","summary":""},{"assertionResults":[{"ancestorTitles":[],"failureMessages":[],"fullName":"renders without crashing","location":null,"status":"passed","title":"renders without crashing"}],"endTime":1676120756947,"message":"","name":"/Users/gokivumbi/Documents/Projects/Else/Dufuna-CodeCamp-21/submissions/kivumbiGeorge/react/react-app/src/__tests__/App.test.js","startTime":1676120756161,"status":"passed","summary":""}],"wasInterrupted":false}
